requirements.txt
# linting:
black
flake8
isort
mamushi
pip-tools
pre-commit

# testing:
eip712
eth_account
ipython
hypothesis
pytest
pytest-xdist
pytest-forked
pytest-repeat
pdbpp
hypothesis>=6.68.1

# analytics
pandas
matplotlib

# vyper and dev framework:
git+https://github.com/DanielSchiavini/titanoboa-zksync
titanoboa@git+https://github.com/DanielSchiavini/titanoboa@zksync
vyper>=0.3.10
